task 4

+++ part3/hbnb/app/services/facade.py
@@ -28,14 +28,22 @@ class HBnBFacade:
     def get_all_users(self):
         return self.user_repo.get_all()
 
-    def update_user(self, user_id, user_data):
+    def update_user(self, req_id, user_id, user_data):
         user = self.get_user(user_id)
         if (not user):
             raise UserNotFoundError
         if (user.email != user_data["email"] and self.get_user_by_email(user_data["email"])):
             raise EmailTakenError
+        if (req_id != user_id):
+            raise CannotUpdateOthersError
         self.user_repo.update(user_id, user_data)
         return self.get_user(user_id)
+
+    def user_owns_place(self, user_id, place_id):
+        place = self.get_place(place_id)
+        if not place or place.owner_id != user_id:
+            return False
+        return True
 
     def create_amenity(self, amenity_data):
         amenity = Amenity(**amenity_data)
@@ -69,9 +77,12 @@ class HBnBFacade:
     def get_all_places(self):
         return self.place_repo.get_all()
 
-    def update_place(self, place_id, place_data):
-        if not self.get_place(place_id):
+    def update_place(self, user_id, place_id, place_data):
+        place = self.get_place(place_id)
+        if not place:
             raise PlaceNotFoundError
+        if user_id != place.owner_id:
+            raise DoesNotOwnPlaceError
         self.place_repo.update(place_id, place_data)
         return self.get_place(place_id)
 
@@ -82,6 +93,10 @@ class HBnBFacade:
             raise UserNotFoundError
         if (not place):
             raise PlaceNotFoundError
+        if (self.user_owns_place(user.id, place.id)):
+            raise CannotReviewOwnPlaceError
+        if (user.id in [facade.get_review(r).user_id for r in place.reviews]):
+            raise AlreadyReviewedError
         review = Review(**review_data)
         user.addReview(review)
         place.addReview(review)
@@ -100,20 +115,23 @@ class HBnBFacade:
             raise PlaceNotFoundError
         return [self.get_review(review_id) for review_id in place.reviews]
 
-    def update_review(self, review_id, review_data):
-        if not self.get_review(review_id):
-            raise ReviewNotFoundError
-        self.review_repo.update(review_id, review_data)
-        return self.get_review(review_id)
-
-    def delete_review(self, review_id):
+    def update_review(self, user_id, review_id, review_data):
         review = self.get_review(review_id)
         if not review:
             raise ReviewNotFoundError
-        place = self.get_place(review.place_id)
-        user = self.get_user(review.user_id)
-        user.removeReview(review)
-        place.removeReview(review)
+        if review.user_id != user_id:
+            raise DoesNotOwnReviewError
+        self.review_repo.update(review_id, review_data)
+        return self.get_review(review_id)
+
+    def delete_review(self, user_id, review_id):
+        review = self.get_review(review_id)
+        if not review:
+            raise ReviewNotFoundError
+        if review.user_id != user_id:
+            raise DoesNotOwnReviewError
+        place = self.get_place(review.place_id).removeReview(review)
+        user = self.get_user(review.user_id).removeReview(review)
         self.review_repo.delete(review_id)
 
     def reset(self):
